+++ git-updater-gitlab.php
@@ -13,7 +13,7 @@
  * Plugin Name:       GitHub Updater - GitLab
  * Plugin URI:        https://github.com/afragen/git-updater-gitlab
  * Description:       Add GitLab hosted repositories to the GitHub Updater plugin.
- * Version:           0.7.0
+ * Version:           0.7.1
  * Author:            Andy Fragen
  * License:           MIT
  * Network:           true