+++ src/Bootstrap.php
@@ -59,7 +59,9 @@ class Bootstrap {
 		add_filter( 'gu_api_url_type', [ $this, 'set_api_url_data' ], 10, 4 );
 		add_filter( 'gu_post_get_credentials', [ $this, 'set_credentials' ], 10, 2 );
 		add_filter( 'gu_get_auth_header', [ $this, 'set_auth_header' ], 10, 2 );
+		add_filter( 'gu_decode_response', [ $this, 'decode_response' ], 10, 2 );
 		add_filter( 'gu_git_servers', [ $this, 'set_git_servers' ], 10, 1 );
+		add_filter( 'gu_running_git_servers', [ $this, 'set_running_enterprise_servers' ], 10, 2 );
 		add_filter( 'gu_installed_apis', [ $this, 'set_installed_apis' ], 10, 1 );
 		add_filter( 'gu_parse_release_asset', [ $this, 'parse_release_asset' ], 10, 4 );
 		add_filter( 'gu_install_remote_install', [ $this, 'set_remote_install_data' ], 10, 2 );
@@ -228,6 +230,23 @@ class Bootstrap {
 	}
 
 	/**
+	 * Decode API response.
+	 *
+	 * @param \stdClass $response API response object.
+	 * @param string    $gitName  of API, eg 'github'.
+	 *
+	 * @return \stdClass
+	 */
+	public function decode_response( $response, $git ) {
+		if ( 'gitlab' === $git ) {
+			// phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions.obfuscation_base64_decode
+			$response = isset( $response->content ) ? base64_decode( $response->content ) : $response;
+		}
+
+		return $response;
+	}
+
+	/**
 	 * Add API as git server.
 	 *
 	 * @param array $git_servers Array of git servers.
@@ -236,6 +255,30 @@ class Bootstrap {
 	 */
 	public function set_git_servers( $git_servers ) {
 		return array_merge( $git_servers, [ 'gitlab' => 'GitLab' ] );
+	}
+
+	/**
+	 * Set running enterprise git servers.
+	 *
+	 * @param array $servers Array of repository git types.
+	 * @param array $repos   Array of repositories objects.
+	 *
+	 * @return array
+	 */
+	public function set_running_enterprise_servers( $servers, $repos ) {
+		$ent_servers = array_map(
+			function ( $e ) {
+				if ( ! empty( $e->enterprise ) ) {
+					if ( 'gitlab' === $e->git ) {
+						return 'gitlabce';
+					}
+				}
+			},
+			$repos
+		);
+		$ent_servers = array_filter( $ent_servers );
+
+		return array_merge( $servers, $ent_servers );
 	}
 
 	/**